Add language-specific constructor refactoring provider implementations
