+++ src/Features/Core/Portable/GenerateConstructorFromMembers/AbstractGenerateConstructorFromMembersCodeRefactoringProvider.ConstructorDelegatingCodeAction.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -9,21 +9,22 @@ using Microsoft.CodeAnalysis.CodeGeneration;
 using Microsoft.CodeAnalysis.CodeStyle;
 using Microsoft.CodeAnalysis.Editing;
 using Microsoft.CodeAnalysis.LanguageServices;
+using Microsoft.CodeAnalysis.PooledObjects;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 
 namespace Microsoft.CodeAnalysis.GenerateConstructorFromMembers
 {
-    internal partial class GenerateConstructorFromMembersCodeRefactoringProvider
+    internal abstract partial class AbstractGenerateConstructorFromMembersCodeRefactoringProvider
     {
-        private class FieldDelegatingCodeAction : CodeAction
+        private class ConstructorDelegatingCodeAction : CodeAction
         {
-            private readonly GenerateConstructorFromMembersCodeRefactoringProvider _service;
+            private readonly AbstractGenerateConstructorFromMembersCodeRefactoringProvider _service;
             private readonly Document _document;
             private readonly State _state;
             private readonly bool _addNullChecks;
 
-            public FieldDelegatingCodeAction(
-                GenerateConstructorFromMembersCodeRefactoringProvider service,
+            public ConstructorDelegatingCodeAction(
+                AbstractGenerateConstructorFromMembersCodeRefactoringProvider service,
                 Document document,
                 State state,
                 bool addNullChecks)
@@ -42,28 +43,38 @@ namespace Microsoft.CodeAnalysis.GenerateConstructorFromMembers
                 //
                 // Otherwise, just generate a normal constructor that assigns any provided
                 // parameters into fields.
-                var parameterToExistingFieldMap = new Dictionary<string, ISymbol>();
-                for (var i = 0; i < _state.Parameters.Length; i++)
-                {
-                    parameterToExistingFieldMap[_state.Parameters[i].Name] = _state.SelectedMembers[i];
-                }
-
-                var factory = _document.GetLanguageService<SyntaxGenerator>();
+                var project = _document.Project;
+                var languageServices = project.Solution.Workspace.Services.GetLanguageServices(_state.ContainingType.Language);
 
                 var semanticModel = await _document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-                var syntaxTree = semanticModel.SyntaxTree;
-                var options = await _document.GetOptionsAsync(cancellationToken).ConfigureAwait(false);
-                var preferThrowExpression = options.GetOption(CodeStyleOptions.PreferThrowExpression).Value;
+                var factory = languageServices.GetService<SyntaxGenerator>();
+                var codeGenerationService = languageServices.GetService<ICodeGenerationService>();
 
-                var (fields, constructor) = factory.CreateFieldDelegatingConstructor(
-                    semanticModel,
-                    _state.ContainingType.Name,
-                    _state.ContainingType,
-                    _state.Parameters,
-                    parameterToExistingFieldMap,
-                    parameterToNewFieldMap: null,
-                    addNullChecks: _addNullChecks,
-                    preferThrowExpression: preferThrowExpression);
+                var thisConstructorArguments = factory.CreateArguments(
+                    _state.Parameters.Take(_state.DelegatedConstructor.Parameters.Length).ToImmutableArray());
+
+                var nullCheckStatements = ArrayBuilder<SyntaxNode>.GetInstance();
+                var assignStatements = ArrayBuilder<SyntaxNode>.GetInstance();
+
+                var options = await _document.GetOptionsAsync(cancellationToken).ConfigureAwait(false);
+                var useThrowExpressions = _service.GetPreferThrowExpressionOptionValue(options);
+
+                for (var i = _state.DelegatedConstructor.Parameters.Length; i < _state.Parameters.Length; i++)
+                {
+                    var symbolName = _state.SelectedMembers[i].Name;
+                    var parameter = _state.Parameters[i];
+
+                    var fieldAccess = factory.MemberAccessExpression(
+                        factory.ThisExpression(),
+                        factory.IdentifierName(symbolName));
+
+                    factory.AddAssignmentStatements(
+                        semanticModel, parameter, fieldAccess,
+                        _addNullChecks, useThrowExpressions,
+                        nullCheckStatements, assignStatements);
+                }
+
+                var syntaxTree = await _document.GetSyntaxTreeAsync(cancellationToken).ConfigureAwait(false);
 
                 // If the user has selected a set of members (i.e. TextSpan is not empty), then we will
                 // choose the right location (i.e. null) to insert the constructor.  However, if they're 
@@ -73,14 +84,22 @@ namespace Microsoft.CodeAnalysis.GenerateConstructorFromMembers
                     ? syntaxTree.GetLocation(_state.TextSpan)
                     : null;
 
-                var result = await CodeGenerator.AddMemberDeclarationsAsync(
+                var statements = nullCheckStatements.ToImmutableAndFree().Concat(assignStatements.ToImmutableAndFree());
+                var result = await codeGenerationService.AddMethodAsync(
                     _document.Project.Solution,
                     _state.ContainingType,
-                    fields.Concat(constructor),
+                    CodeGenerationSymbolFactory.CreateConstructorSymbol(
+                        attributes: default,
+                        accessibility: Accessibility.Public,
+                        modifiers: new DeclarationModifiers(),
+                        typeName: _state.ContainingType.Name,
+                        parameters: _state.Parameters,
+                        statements: statements,
+                        thisConstructorArguments: thisConstructorArguments),
                     new CodeGenerationOptions(
                         contextLocation: syntaxTree.GetLocation(_state.TextSpan),
                         afterThisLocation: afterThisLocation),
-                    cancellationToken).ConfigureAwait(false);
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
 
                 return await AddNavigationAnnotationAsync(result, cancellationToken).ConfigureAwait(false);
             }
@@ -93,16 +112,8 @@ namespace Microsoft.CodeAnalysis.GenerateConstructorFromMembers
                     var parameters = _state.Parameters.Select(p => symbolDisplayService.ToDisplayString(p, SimpleFormat));
                     var parameterString = string.Join(", ", parameters);
 
-                    if (_state.DelegatedConstructor == null)
-                    {
-                        return string.Format(FeaturesResources.Generate_constructor_0_1,
-                            _state.ContainingType.Name, parameterString);
-                    }
-                    else
-                    {
-                        return string.Format(FeaturesResources.Generate_field_assigning_constructor_0_1,
-                            _state.ContainingType.Name, parameterString);
-                    }
+                    return string.Format(FeaturesResources.Generate_delegating_constructor_0_1,
+                        _state.ContainingType.Name, parameterString);
                 }
             }
         }
